+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 * Introduced separate settinge page and reduced widget backview to widget settings only
+* Add options to track logged in users, feeds and search requests
 
 ## 1.6.3
 * Fix compatibility issue with some PHP implementations not populating `INPUT_SERVER`